+++ core/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [Unreleased]
 
-- ???
+- CaptureWebSocket serves webapp index.html with bootstrap metadata
 
 ## [v0.0.5] - 2025-06-12